Move validation routines to separate submodule
